+++ examples/snake/src/lib/degreesToRad.ts
@@ -1,0 +1,3 @@
+export function degreesToRad(degrees: number) {
+  return degrees * (Math.PI / 180)
+}
