+++ examples/snake/src/lib/getRandomIntBetween.ts
@@ -1,0 +1,3 @@
+export function getRandomIntBetween(min: number, max: number) {
+  return Math.floor(Math.random() * (max - min) + min)
+}
